Extract simple inlining utility from ArrayConstructorLowering
